Moving the node and authservice common code to the nextprotocol class
